+++ miniOgre/particle/ParticleFX/OgreTexture1PosInterpolatorAffector.cpp
@@ -105,7 +105,7 @@ namespace Ogre {
 					auto mat = MaterialManager::getSingleton().getByName(name);
 					if (mat)
 					{
-						auto texUnit = mat->getTextureUnit(0);
+						auto texUnit = mat->getTextureUnit((uint32_t)0);
 						texUnit->setTextureScroll(u, v);
 					}
 				}
