+++ miniOgre/particle/ParticleFX/OgreTexture1RotateInterpolatorAffector.cpp
@@ -101,7 +101,7 @@ namespace Ogre {
 					auto mat = MaterialManager::getSingleton().getByName(name);
 					if (mat)
 					{
-						auto texUnit = mat->getTextureUnit(0);
+						auto texUnit = mat->getTextureUnit((uint32_t)0);
 						Ogre::Degree degree(rotate);
 						Ogre::Radian f(degree);
 						texUnit->setTextureRotate(f);
